move formatDate to util
