+++ client/wfnews-war/src/main/angular/src/app/components/preview-panels/danger-rating-preview/danger-rating-preview.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { ResourcesRoutes } from '@app/utils';
+import { ResourcesRoutes, formatDate } from '@app/utils';
 
 @Component({
   selector: 'wfnews-danger-rating-preview',
@@ -14,6 +14,7 @@ export class DangerRatingPreviewComponent {
   ) {}
 
   public data;
+  formatDate = formatDate
   setContent(data) {
     this.data = data.properties;
   }
@@ -36,19 +37,6 @@ export class DangerRatingPreviewComponent {
     (
       document.getElementsByClassName('desktop-preview').item(0) as HTMLElement
     ).style.display = 'none';
-  }
-  
-  formatDate(timestamp: string | number): string {
-    if (timestamp) {
-      const date = new Date((typeof timestamp === 'string' ? timestamp.slice(0, 10) : timestamp));
-      const options: Intl.DateTimeFormatOptions = {
-        year: 'numeric',
-        month: 'long',
-        day: 'numeric',
-      };
-
-      return date.toLocaleDateString('en-US', options);
-    } else return '';
   }
 
   enterFullDetail(){
