+++ client/wfnews-war/src/main/angular/src/app/components/preview-panels/protected-land-preview/protected-land-preview.component.html
@@ -18,10 +18,14 @@
                 <span>Effective on {{formatDate(data.ACCESS_STATUS_EFFECTIVE_DATE)}}</span>
             </div>
             <div class="identify-card-item">
-                <img *ngIf="data.CAMPFIRE_BAN_IND !== 'N'" alt="icon" class="incident-icon" src="/assets/images/svg-icons/campfire-allowed.svg">
-                <img *ngIf="data.CAMPFIRE_BAN_IND == 'N'" alt="icon" class="incident-icon" src="/assets/images/svg-icons/ban.svg">
-                <span *ngIf="data.CAMPFIRE_BAN_IND !== 'N'">Campfire Allowed</span>
-                <span *ngIf="data.CAMPFIRE_BAN_IND == 'N'">Campfire Ban</span>
+                <img
+                  alt="icon"
+                  class="incident-icon"
+                  [src]="data.CAMPFIRE_BAN_IND === 'N' ? '/assets/images/svg-icons/ban.svg' : '/assets/images/svg-icons/campfire-allowed.svg'"
+                >
+                <span>
+                  {{ data.CAMPFIRE_BAN_IND === 'N' ? 'Campfire Ban' : 'Campfire Allowed' }}
+                </span>
             </div>
             <div class="identify-card-item">
                 <span>{{data.ACCESS_DETAILS ? data.ACCESS_DETAILS : 'This park is currently closed due to wildfire.'}}</span>
